appropriate wait times
