+++ tools/integration_tests/util/operations/file_operations.go
@@ -33,6 +33,7 @@ import (
 	"time"
 
 	"github.com/googlecloudplatform/gcsfuse/v3/internal/storage/gcs"
+	"github.com/googlecloudplatform/gcsfuse/v3/tools/integration_tests/util/setup"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/context"
@@ -185,6 +186,7 @@ func CloseFile(file *os.File) {
 	if err := file.Close(); err != nil {
 		log.Fatalf("error in closing: %v", err)
 	}
+	WaitForSizeUpdate(setup.IsZonalBucketRun(), 5*time.Second)
 }
 
 func RemoveFile(filePath string) {
@@ -620,6 +622,7 @@ func WriteAt(content string, offset int64, fh *os.File, t testing.TB) {
 func CloseFileShouldNotThrowError(t testing.TB, file *os.File) {
 	err := file.Close()
 	assert.NoError(t, err)
+	WaitForSizeUpdate(setup.IsZonalBucketRun(), 5*time.Second)
 }
 
 func CloseFileShouldThrowError(t *testing.T, file *os.File) {
@@ -636,6 +639,7 @@ func SyncFile(fh *os.File, t *testing.T) {
 	if err != nil {
 		t.Fatalf("%s.Sync(): %v", fh.Name(), err)
 	}
+	WaitForSizeUpdate(setup.IsZonalBucketRun(), time.Minute)
 }
 
 func SyncFileShouldThrowError(t *testing.T, file *os.File) {
@@ -845,5 +849,25 @@ func ValidateCloseGivenThatFileIsRenamed(t *testing.T, file *os.File, streamingW
 		require.NoError(t, err)
 	} else {
 		ValidateESTALEError(t, err)
+	}
+}
+
+// CreateFileAndCopyToMntDir creates a file of given size.
+// The same file will be copied to the mounted directory as well.
+func CreateFileAndCopyToMntDir(t *testing.T, fileSize int, dirName string) (string, string) {
+	testDir := setup.SetupTestDirectory(dirName)
+	fileInLocalDisk := "test_file" + setup.GenerateRandomString(5) + ".txt"
+	filePathInLocalDisk := path.Join(os.TempDir(), fileInLocalDisk)
+	filePathInMntDir := path.Join(testDir, fileInLocalDisk)
+	CreateFileOnDiskAndCopyToMntDir(t, filePathInLocalDisk, filePathInMntDir, fileSize)
+	return filePathInLocalDisk, filePathInMntDir
+}
+
+// CreateFileOnDiskAndCopyToMntDir creates a file of given size and copies to given path.
+func CreateFileOnDiskAndCopyToMntDir(t *testing.T, filePathInLocalDisk string, filePathInMntDir string, fileSize int) {
+	setup.RunScriptForTestData("../util/setup/testdata/write_content_of_fix_size_in_file.sh", filePathInLocalDisk, strconv.Itoa(fileSize))
+	err := CopyFile(filePathInLocalDisk, filePathInMntDir)
+	if err != nil {
+		t.Errorf("Error in copying file:%v", err)
 	}
 }
